+++ packages/common_client/lib/src/context_modifiers/env_context_modifier.dart
@@ -42,6 +42,11 @@ final class AutoEnvContextModifier implements ContextModifier {
 
   @override
   Future<LDContext> decorate(LDContext context) async {
+    if (!context.valid) {
+      _logger.warn(
+          'AutoEnvContextModifier was asked to modify an invalid context and will attempt to do so. This is expected if starting with an empty context.');
+    }
+
     final builder = LDContextBuilder.fromContext(context);
 
     for (final recipe in _recipes) {