+++ packages/event_source_client/lib/src/sse_client_html.dart
@@ -16,6 +16,8 @@ class HtmlSseClient implements SSEClient {
   /// This controller is for the events going to the subscribers of this client.
   late final StreamController<ld_message_event.Event> _messageEventsController;
 
+  late final EventSourceLogger _logger;
+
   Backoff _backoff = Backoff(math.Random());
 
   final Uri _uri;
@@ -26,9 +28,10 @@ class HtmlSseClient implements SSEClient {
 
   /// Creates an instance of an SSEClient that will connect in the future
   /// to the [uri] provided.
-  HtmlSseClient(Uri uri, Set<String> eventTypes)
+  HtmlSseClient(Uri uri, Set<String> eventTypes, EventSourceLogger? logger)
       : _uri = uri,
         _eventTypes = eventTypes {
+    _logger = logger ?? NoOpLogger();
     _messageEventsController =
         StreamController<ld_message_event.Event>.broadcast(
       onListen: () {
@@ -90,7 +93,10 @@ class HtmlSseClient implements SSEClient {
   Stream<Event> get stream => _messageEventsController.stream;
 
   @override
-  Future close() => _messageEventsController.close();
+  Future close() async {
+    _logger.debug('Closing SSE client permanently.');
+    _messageEventsController.close();
+  }
 
   @override
   void restart() {
@@ -115,6 +121,7 @@ SSEClient getSSEClient(
         Duration connectTimeout,
         Duration readTimeout,
         String? body,
-        String method) =>
+        String method,
+        EventSourceLogger? logger) =>
     // dropping unsupported configuration options
-    HtmlSseClient(uri, eventTypes);
+    HtmlSseClient(uri, eventTypes, logger);